pyproject.toml
[project]
name        = "pianola"
description = "🎹 Smoothest piano roll visualizer"
authors     = [{name="Tremeschin"}]
dynamic     = ["keywords"]
version     = "0.10.0"
readme      = "readme.md"
license     = "AGPL-3.0"
requires-python = ">=3.10"
dependencies = [
    "dearlog",
    "pooch~=1.0",
    "shaderflow",
    "music21>=9.0",
    "Pillow",
]

[project.scripts]
pianola = "pianola.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
